Transform `ethereum.ts` file to `lib/ethereum` shared lib
