+++ typescript/src/lib/ethereum/contract-handle.ts
@@ -1,15 +1,145 @@
+import { providers, Signer, utils } from "ethers"
 import {
   Contract as EthersContract,
   Event as EthersEvent,
   EventFilter as EthersEventFilter,
-} from "ethers"
+} from "@ethersproject/contracts"
+import { GetEvents } from "../contracts"
 import {
   backoffRetrier,
-  skipRetryWhenMatched,
   ExecutionLoggerFn,
-} from "./backoff"
+  skipRetryWhenMatched,
+} from "../../backoff"
+import { Address } from "./address"
 
-const GET_EVENTS_BLOCK_INTERVAL = 10_000
+/**
+ * Contract deployment artifact.
+ * @see [hardhat-deploy#Deployment](https://github.com/wighawag/hardhat-deploy/blob/0c969e9a27b4eeff9f5ccac7e19721ef2329eed2/types.ts#L358)}
+ */
+export interface Deployment {
+  /**
+   * Address of the deployed contract.
+   */
+  address: string
+  /**
+   * Contract's ABI.
+   */
+  abi: any[]
+  /**
+   * Deployment transaction receipt.
+   */
+  receipt: {
+    /**
+     * Number of block in which the contract was deployed.
+     */
+    blockNumber: number
+  }
+}
+
+/**
+ * Represents a config set required to connect an Ethereum contract.
+ */
+export interface ContractConfig {
+  /**
+   * Address of the Ethereum contract as a 0x-prefixed hex string.
+   * Optional parameter, if not provided the value will be resolved from the
+   * contract artifact.
+   */
+  address?: string
+  /**
+   * Signer - will return a Contract which will act on behalf of that signer. The signer will sign all contract transactions.
+   * Provider - will return a downgraded Contract which only has read-only access (i.e. constant calls)
+   */
+  signerOrProvider: Signer | providers.Provider
+  /**
+   * Number of a block in which the contract was deployed.
+   * Optional parameter, if not provided the value will be resolved from the
+   * contract artifact.
+   */
+  deployedAtBlockNumber?: number
+}
+
+/**
+ * Deployed Ethereum contract
+ */
+export class EthereumContract<T extends EthersContract> {
+  /**
+   * Ethers instance of the deployed contract.
+   */
+  protected readonly _instance: T
+  /**
+   * Number of a block within which the contract was deployed. Value is read from
+   * the contract deployment artifact. It can be overwritten by setting a
+   * {@link ContractConfig.deployedAtBlockNumber} property.
+   */
+  protected readonly _deployedAtBlockNumber: number
+  /**
+   * Number of retries for ethereum requests.
+   */
+  protected readonly _totalRetryAttempts: number
+
+  /**
+   * @param config Configuration for contract instance initialization.
+   * @param deployment Contract Deployment artifact.
+   * @param totalRetryAttempts Number of retries for ethereum requests.
+   */
+  constructor(
+    config: ContractConfig,
+    deployment: Deployment,
+    totalRetryAttempts = 3
+  ) {
+    this._instance = new EthersContract(
+      config.address ?? utils.getAddress(deployment.address),
+      `${JSON.stringify(deployment.abi)}`,
+      config.signerOrProvider
+    ) as T
+
+    this._deployedAtBlockNumber =
+      config.deployedAtBlockNumber ?? deployment.receipt.blockNumber
+
+    this._totalRetryAttempts = totalRetryAttempts
+  }
+
+  /**
+   * Get address of the contract instance.
+   * @returns Address of this contract instance.
+   */
+  getAddress(): Address {
+    return Address.from(this._instance.address)
+  }
+
+  /**
+   * Get events emitted by the Ethereum contract.
+   * It starts searching from provided block number. If the {@link GetEvents.Options#fromBlock}
+   * option is missing it looks for a contract's defined property
+   * {@link _deployedAtBlockNumber}. If the property is missing starts searching
+   * from block `0`.
+   * @param eventName Name of the event.
+   * @param options Options for events fetching.
+   * @param filterArgs Arguments for events filtering.
+   * @returns Array of found events.
+   */
+  async getEvents(
+    eventName: string,
+    options?: GetEvents.Options,
+    ...filterArgs: Array<unknown>
+  ): Promise<EthersEvent[]> {
+    // TODO: Test if we need a workaround for querying events from big range in chunks,
+    // see: https://github.com/keep-network/tbtc-monitoring/blob/e169357d7b8c638d4eaf73d52aa8f53ee4aebc1d/src/lib/ethereum-helper.js#L44-L73
+    return backoffRetrier<EthersEvent[]>(
+      options?.retries ?? this._totalRetryAttempts
+    )(async () => {
+      return await getEvents(
+        this._instance,
+        this._instance.filters[eventName](...filterArgs),
+        options?.fromBlock ?? this._deployedAtBlockNumber,
+        options?.toBlock,
+        options?.batchedQueryBlockInterval,
+        options?.logger
+      )
+    })
+  }
+}
 
 /**
  * Sends ethereum transaction with retries.
@@ -80,6 +210,8 @@ function resolveEthersError(err: unknown): unknown {
   return err
 }
 
+const GET_EVENTS_BLOCK_INTERVAL = 10_000
+
 /**
  * Looks up all existing events defined by the {@link event} filter on
  * {@link sourceContract}, searching past blocks and then returning them.
@@ -97,7 +229,7 @@ function resolveEthersError(err: unknown): unknown {
  * @returns A promise that will be fulfilled by the list of event objects once
  *          they are found.
  */
-export async function getEvents(
+async function getEvents(
   sourceContract: EthersContract,
   event: EthersEventFilter,
   fromBlock: number = 0,
