+++ typescript/src/lib/ethereum/index.ts
@@ -1,0 +1,10 @@
+export * from "./address"
+export * from "./bridge"
+export * from "./tbtc-token"
+export * from "./tbtc-vault"
+export * from "./wallet-registry"
+
+// The `contract-handle` module should not be re-exported directly as it
+// contains low-level contract integration code. Re-export only components
+// that are relevant for `lib/ethereum` clients.
+export { ContractConfig } from "./contract-handle"
