+++ typescript/test/data/redemption.ts
@@ -10,8 +10,8 @@ import {
   createOutputScriptFromAddress,
 } from "../../src/lib/bitcoin"
 import { RedemptionRequest } from "../../src/redemption"
-import { Address } from "../../src/ethereum"
-import { BitcoinNetwork, BitcoinTransaction, Hex } from "../../src"
+import { Address } from "../../src/lib/ethereum"
+import { BitcoinTransaction, Hex } from "../../src"
 import { WalletState } from "../../src/wallet"
 
 /**
@@ -723,8 +723,7 @@ export const findWalletForRedemptionData: {
               outputIndex: 0,
               value: BigNumber.from("791613461"),
               scriptPubKey: createOutputScriptFromAddress(
-                "tb1qqwm566yn44rdlhgph8sw8vecta8uutg79afuja",
-                BitcoinNetwork.Testnet
+                "tb1qqwm566yn44rdlhgph8sw8vecta8uutg79afuja"
               ),
             },
           ],
@@ -849,8 +848,7 @@ export const findWalletForRedemptionData: {
               outputIndex: 0,
               value: BigNumber.from("3370000"), // 0.0337 BTC
               scriptPubKey: createOutputScriptFromAddress(
-                "tb1qx2xejtjltdcau5dpks8ucszkhxdg3fj88404lh",
-                BitcoinNetwork.Testnet
+                "tb1qx2xejtjltdcau5dpks8ucszkhxdg3fj88404lh"
               ),
             },
           ],
